Move request types to separate files
