+++ rosetta/request/combine.go
@@ -1,0 +1,28 @@
+// Copyright 2021 Optakt Labs OÜ
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy of
+// the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations under
+// the License.
+
+package request
+
+import (
+	"github.com/optakt/flow-dps/rosetta/identifier"
+	"github.com/optakt/flow-dps/rosetta/object"
+)
+
+// Combine implements the request schema for /construction/combine.
+// See https://www.rosetta-api.org/docs/ConstructionApi.html#request
+type Combine struct {
+	NetworkID           identifier.Network `json:"network_identifier"`
+	UnsignedTransaction string             `json:"unsigned_transaction"`
+	Signatures          []object.Signature `json:"signatures"`
+}
