--- rosetta/request/request.go
@@ -1,115 +1,0 @@
-// Copyright 2021 Optakt Labs OÜ
-//
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not
-// use this file except in compliance with the License. You may obtain a copy of
-// the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations under
-// the License.
-
-package request
-
-import (
-	"github.com/optakt/flow-dps/rosetta/identifier"
-	"github.com/optakt/flow-dps/rosetta/object"
-)
-
-// Balance implements the request schema for /account/balance.
-// See https://www.rosetta-api.org/docs/AccountApi.html#request
-type Balance struct {
-	NetworkID  identifier.Network    `json:"network_identifier"`
-	BlockID    identifier.Block      `json:"block_identifier"`
-	AccountID  identifier.Account    `json:"account_identifier"`
-	Currencies []identifier.Currency `json:"currencies"`
-}
-
-// Block implements the request schema for /block.
-// See https://www.rosetta-api.org/docs/BlockApi.html#request
-type Block struct {
-	NetworkID identifier.Network `json:"network_identifier"`
-	BlockID   identifier.Block   `json:"block_identifier"`
-}
-
-// Combine implements the request schema for /construction/combine.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request
-type Combine struct {
-	NetworkID           identifier.Network `json:"network_identifier"`
-	UnsignedTransaction string             `json:"unsigned_transaction"`
-	Signatures          []object.Signature `json:"signatures"`
-}
-
-// Hash implements the request schema for /construction/hash.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-2
-type Hash struct {
-	NetworkID         identifier.Network `json:"network_identifier"`
-	SignedTransaction string             `json:"signed_transaction"`
-}
-
-// Metadata implements the request schema for /construction/metadata.
-// `Options` object in this request is generated by a call to `/construction/preprocess`,
-// and should be sent unaltered as returned by that endpoint.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-3
-type Metadata struct {
-	NetworkID identifier.Network `json:"network_identifier"`
-	Options   object.Options     `json:"options"`
-}
-
-// Networks implements the empty request schema for the /network/list endpoint.
-// See https://www.rosetta-api.org/docs/NetworkApi.html#request
-type Networks struct {
-}
-
-// Options implements the empty request schema for the /network/options endpoint.
-// See https://www.rosetta-api.org/docs/NetworkApi.html#request-1
-type Options struct {
-	NetworkID identifier.Network `json:"network_identifier"`
-}
-
-// Parse implements the request schema for /construction/parse.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-4
-type Parse struct {
-	NetworkID   identifier.Network `json:"network_identifier"`
-	Signed      bool               `json:"signed"`
-	Transaction string             `json:"transaction"`
-}
-
-// Payloads implements the request schema for /construction/payloads.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-5
-type Payloads struct {
-	NetworkID  identifier.Network `json:"network_identifier"`
-	Operations []object.Operation `json:"operations"`
-	Metadata   object.Metadata    `json:"metadata"`
-}
-
-// Preprocess implements the request schema for /construction/preprocess.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-6
-type Preprocess struct {
-	NetworkID  identifier.Network `json:"network_identifier"`
-	Operations []object.Operation `json:"operations"`
-}
-
-// Status implements the request schema for /network/status.
-// See https://www.rosetta-api.org/docs/NetworkApi.html#request-2
-type Status struct {
-	NetworkID identifier.Network `json:"network_identifier"`
-}
-
-// Submit implements the request schema for /construction/submit.
-// See https://www.rosetta-api.org/docs/ConstructionApi.html#request-7
-type Submit struct {
-	NetworkID         identifier.Network `json:"network_identifier"`
-	SignedTransaction string             `json:"signed_transaction"`
-}
-
-// Transaction implements the request schema for /block/transaction.
-// See https://www.rosetta-api.org/docs/BlockApi.html#request-1
-type Transaction struct {
-	NetworkID     identifier.Network     `json:"network_identifier"`
-	BlockID       identifier.Block       `json:"block_identifier"`
-	TransactionID identifier.Transaction `json:"transaction_identifier"`
-}
